refactor: 디렉토리 위치 변경 등
