--- src/main/java/com/zerobase/funding/api/auth/dto/LogoutResponse.java
@@ -1,5 +1,0 @@
-package com.zerobase.funding.api.auth.dto;
-
-public record LogoutResponse(String memberKey) {
-
-}
