+++ src/main/java/com/zerobase/funding/api/auth/dto/OAuth2UserInfo.java
@@ -1,35 +1,26 @@
 package com.zerobase.funding.api.auth.dto;
 
-import static com.zerobase.funding.api.exception.ErrorCode.ILLEGAL_REGISTRATION_ID;
+import static com.zerobase.funding.global.exception.ErrorCode.ILLEGAL_REGISTRATION_ID;
 
 import com.zerobase.funding.api.auth.exception.AuthException;
+import com.zerobase.funding.api.utils.KeyGenerator;
 import com.zerobase.funding.domain.member.entity.Member;
 import com.zerobase.funding.domain.member.entity.Role;
-import com.zerobase.funding.global.utils.KeyGenerator;
 import java.util.Map;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
 @Builder
-public class OAuth2UserInfo {
-
-    private String name;
-    private String email;
-    private String profile;
+public record OAuth2UserInfo(
+        String name,
+        String email,
+        String profile
+) {
 
     public static OAuth2UserInfo of(String registrationId, Map<String, Object> attributes) {
         return switch (registrationId) {
-            case "google" ->
-                    ofGoogle(attributes);
-            case "kakao" ->
-                    ofKakao(attributes);
-            default ->
-                    throw new AuthException(ILLEGAL_REGISTRATION_ID);
+            case "google" -> ofGoogle(attributes);
+            case "kakao" -> ofKakao(attributes);
+            default -> throw new AuthException(ILLEGAL_REGISTRATION_ID);
         };
     }
