+++ src/main/java/com/zerobase/funding/api/auth/exception/AuthException.java
@@ -1,7 +1,7 @@
 package com.zerobase.funding.api.auth.exception;
 
-import com.zerobase.funding.api.exception.CustomException;
-import com.zerobase.funding.api.exception.ErrorCode;
+import com.zerobase.funding.global.exception.CustomException;
+import com.zerobase.funding.global.exception.ErrorCode;
 
 public class AuthException extends CustomException {
