+++ src/main/java/com/zerobase/funding/api/auth/service/AuthenticationService.java
@@ -1,6 +1,6 @@
 package com.zerobase.funding.api.auth.service;
 
-import static com.zerobase.funding.api.exception.ErrorCode.MEMBER_NOT_FOUND;
+import static com.zerobase.funding.global.exception.ErrorCode.MEMBER_NOT_FOUND;
 
 import com.zerobase.funding.api.member.exception.MemberException;
 import com.zerobase.funding.domain.member.entity.Member;
