+++ src/main/java/com/zerobase/funding/api/converter/ParamTypeConverterFactory.java
@@ -1,4 +1,4 @@
-package com.zerobase.funding.global.converter;
+package com.zerobase.funding.api.converter;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.convert.converter.Converter;
