+++ src/main/java/com/zerobase/funding/api/fundingproduct/exception/FundingProductException.java
@@ -1,7 +1,7 @@
 package com.zerobase.funding.api.fundingproduct.exception;
 
-import com.zerobase.funding.api.exception.CustomException;
-import com.zerobase.funding.api.exception.ErrorCode;
+import com.zerobase.funding.global.exception.CustomException;
+import com.zerobase.funding.global.exception.ErrorCode;
 
 public class FundingProductException extends CustomException {
