+++ src/main/java/com/zerobase/funding/api/fundingproduct/service/FundingProductService.java
@@ -9,7 +9,7 @@ import com.zerobase.funding.domain.fundingproduct.entity.Image;
 import com.zerobase.funding.domain.fundingproduct.entity.ImageType;
 import com.zerobase.funding.domain.fundingproduct.repository.FundingProductRepository;
 import com.zerobase.funding.domain.member.entity.Member;
-import com.zerobase.funding.global.utils.s3.AwsS3Utils;
+import com.zerobase.funding.api.utils.s3.AwsS3Utils;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
