+++ src/main/java/com/zerobase/funding/api/member/service/MemberService.java
@@ -1,6 +1,6 @@
 package com.zerobase.funding.api.member.service;
 
-import static com.zerobase.funding.api.exception.ErrorCode.MEMBER_NOT_FOUND;
+import static com.zerobase.funding.global.exception.ErrorCode.MEMBER_NOT_FOUND;
 
 import com.zerobase.funding.api.member.dto.model.MemberDto;
 import com.zerobase.funding.api.member.dto.MemberEditRequest;
