+++ src/main/java/com/zerobase/funding/global/exception/CustomException.java
@@ -1,4 +1,4 @@
-package com.zerobase.funding.api.exception;
+package com.zerobase.funding.global.exception;
 
 import lombok.Getter;
