+++ src/test/java/com/zerobase/funding/api/auth/controller/AuthControllerTest.java
@@ -2,10 +2,9 @@ package com.zerobase.funding.api.auth.controller;
 
 import static com.zerobase.funding.common.constants.MemberConstants.MEMBER_KEY;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.doNothing;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.zerobase.funding.api.auth.jwt.TokenProvider;
@@ -48,14 +47,12 @@ class AuthControllerTest {
     @DisplayName("로그아웃 처리 - refresh 토큰 삭제")
     void logout() throws Exception {
         // given
-        given(refreshTokenService.deleteRefreshToken(any()))
-                .willReturn(MEMBER_KEY);
+        doNothing().when(refreshTokenService).deleteRefreshToken(any());
 
         // when
         // then
         mockMvc.perform(delete("/auth/logout"))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.memberKey").value(MEMBER_KEY))
+                .andExpect(status().isNoContent())
                 .andDo(print());
     }
 }
